+++ xmake/modules/private/check/checkers/clang/tidy.lua
