+++ xmake/modules/private/check/checkers/cuda/devlink.lua
